litscout/server/requirements.txt
colorama
openai
psycopg2-binary
requests
tqdm